+++ CHANGELOG.md
@@ -16,8 +16,8 @@
 - Console Command class generation with extender.
 - Service Provider class generation with extender.
 - Integration Test class generation.
-
 - Frontend Model generation with its definition.
+- Frontend Modal component generation.
 
 ### Changed